GreenHub/momoapi/requirements.txt
django
djangorestframework

# psychopg2
flask 
python-dotenv
